fix(ci): python-coverage-shards (8) — split shared-base subsystem out of auth_mounts_claude to restore <1500-line maintainability limit
Extract the host-content signature hash and shared read-only overlay-base
build/reap (_host_claude_signature, _shared_claude_base_dir, _chown_tree,
_ensure_shared_claude_base, _reap_stale_claude_base_staging,
_claude_base_staging_build_is_live) plus their constants into a new
auth_mounts_claude_base module, re-imported back so awf.node.auth_mounts.<name>
stays the stable import surface. auth_mounts_claude drops from 1538 to ~1214
lines, under the MAX_FIRST_PARTY_FILE_LINES gate enforced by
test_first_party_code_files_stay_under_line_limit.
